# docs/openapi.yaml
openapi: 3.0.3
info:
  title: "ASE Project API"
  version: "1.0.0"
  description: "API documentation for ASE Project"
servers:
  - url: "http://localhost:5050"
paths:
...
  /auctionbid:
    get:
      summary: "Get all auction bids"
      tags: ["AuctionBid"]
      responses:
        '200':
          description: "List of all auction bids"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AuctionBid'
        '404':
          description: "Auction bids not found"
    post:
      summary: "Create a new auction bid"
      tags: ["AuctionBid"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAuctionBidRequest'
      responses:
        '200':
          description: "Auction bid created successfully"
          content:
            application/json:
              schema:
                type: object
                properties:
                  bidId:
                    type: integer
                    description: "ID of the newly created auction bid"
        '400':
          description: "Invalid auction bid data"
...
  /auction/{status}:
    get:
      summary: "Get all auctions with a specific status"
      tags: ["Auction"]
      parameters:
        - in: path
          name: status
          required: true
          schema:
            type: string
          description: "Status of the auctions to retrieve"
      responses:
        '200':
          description: "List of auctions with the specified status"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Auction'
        '404':
          description: "Auctions not found"
  /auction/{auctionId}:
    get:
      summary: "Get a single auction by ID"
      tags: ["Auction"]
      parameters:
        - in: path
          name: auctionId
          required: true
          schema:
            type: integer
          description: "ID of the auction"
      responses:
        '200':
          description: "Auction retrieved successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Auction'
        '404':
          description: "Auction not found"
    put:
      summary: "Update an auction by ID"
      tags: ["Auction"]
      parameters:
        - in: path
          name: auctionId
          required: true
          schema:
            type: integer
          description: "ID of the auction to update"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAuctionRequest'
      responses:
        '200':
          description: "Auction updated successfully"
        '400':
          description: "Invalid auction data"
        '404':
          description: "Auction not found"
    delete:
      summary: "Delete an auction by ID"
      tags: ["Auction"]
      parameters:
        - in: path
          name: auctionId
          required: true
          schema:
            type: integer
          description: "ID of the auction to delete"
      responses:
        '200':
          description: "Auction deleted successfully"
        '404':
          description: "Auction not found"
  /auction:
    post:
      summary: "Create a new auction"
      tags: ["Auction"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAuctionRequest'
      responses:
        '200':
          description: "Auction created successfully"
          content:
            application/json:
              schema:
                type: object
                properties:
                  auctionId:
                    type: integer
                    description: "ID of the newly created auction"
        '400':
          description: "Invalid auction data" 
...
  /gachacollection/{userId}:
    get:
      summary: "Get all gacha collections for a user"
      tags: ["GachaCollection"]
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
          description: "ID of the user"
      responses:
        '200':
          description: "List of gacha collections for the specified user"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GachaCollection'
        '404':
          description: "Gacha collections not found"
...
  /gachacollection/{collectionId}:
    put:
      summary: "Update a gacha collection by ID"
      tags: ["GachaCollection"]
      parameters:
        - in: path
          name: collectionId
          required: true
          schema:
            type: integer
          description: "ID of the gacha collection to update"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGachaCollectionRequest'
      responses:
        '200':
          description: "Gacha collection updated successfully"
        '400':
          description: "Invalid gacha collection data"
        '404':
          description: "Gacha collection not found"
    delete:
      summary: "Delete a gacha collection by ID"
      tags: ["GachaCollection"]
      parameters:
        - in: path
          name: collectionId
          required: true
          schema:
            type: integer
          description: "ID of the gacha collection to delete"
      responses:
        '200':
          description: "Gacha collection deleted successfully"
        '404':
          description: "Gacha collection not found"    
...
  /player/auction/market:
    get:
      summary: Get active auctions
      description: Retrieve all active auctions.
      tags: [AuctionMicroservice]
      responses:
        "200":
          description: List of active auctions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuctionItem"
        "500":
          description: Internal Server Error
  /player/auction/create:
    post:
      summary: Create a new auction
      description: Create a new auction listing for a gacha item.
      tags:
        - Player
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                gachaCollectionId:
                  type: integer
                auctionStart:
                  type: string
                  format: date-time
                auctionEnd:
                  type: string
                  format: date-time
                minimumBid:
                  type: number
              required:
                - gachaCollectionId
                - auctionStart
                - auctionEnd
                - minimumBid
      responses:
        "201":
          description: Auction created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  auctionId:
                    type: integer
        "400":
          description: Invalid input data
        "500":
          description: Internal Server Error
  /player/auction/bid/{auction_id}:
    post:
      summary: Place a bid on an auction
      description: Place a bid on an active auction.
      tags:
        - Player
      parameters:
        - name: auction_id
          in: path
          required: true
          description: ID of the auction
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bidAmount:
                  type: number
              required:
                - bidAmount
      responses:
        "200":
          description: Bid placed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  bidId:
                    type: integer
        "400":
          description: Invalid request
        "500":
          description: Internal Server Error
  /admin/auction:
    get:
      summary: Get all auctions
      description: Retrieve all auctions.
      tags:
        - Admin
      responses:
        "200":
          description: List of active auctions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuctionItem"
        "500":
          description: Internal Server Error
  /admin/auction/{auction_id}:
    get:
      summary: Get specific auction
      description: Retrieve details of a specific auction.
      tags:
        - Admin
      parameters:
        - name: auction_id
          in: path
          required: true
          description: ID of the auction
          schema:
            type: integer
      responses:
        "200":
          description: Auction details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuctionItem"
        "500":
          description: Internal Server Error
    put:
      summary: Update an auction
      description: Modify details of a specific auction.
      tags:
        - Admin
      parameters:
        - name: auction_id
          in: path
          required: true
          description: ID of the auction
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                auctionStart:
                  type: string
                  format: date-time
                auctionEnd:
                  type: string
                  format: date-time
                gachaCollectionId:
                  type: integer
                minimumBid:
                  type: number
              required:
                - auctionStart
                - auctionEnd
      responses:
        "200":
          description: Auction updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuctionItem"
        "400":
          description: Invalid auction data
        "500":
          description: Internal Server Error
  /admin/auction/history:
    get:
      summary: Get auction history
      description: Retrieve market history of auctions.
      tags:
        - Admin
      responses:
        "200":
          description: Auction history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuctionItem"
        "500":
          description: Internal Server Error
  /admin/auction/history/{user_id}:
    get:
      summary: Get user's auction history
      description: Retrieve the auction history for a specific user.
      tags:
        - Admin
      parameters:
        - name: user_id
          in: path
          required: true
          description: ID of the user
          schema:
            type: integer
      responses:
        "200":
          description: User's auction history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuctionItem"
        "500":
          description: Internal Server Error
components:  
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateAuctionBidRequest:
      type: object
      properties:
        userId:
          type: integer
          description: "ID of the user placing the bid"
        auctionId:
          type: integer
          description: "ID of the auction"
        bidAmount:
          type: number
          format: float
          description: "Amount of the bid"
        timestamp:
          type: string
          format: date-time
          description: "Timestamp of when the bid is placed"
...
    Auction:
      type: object
      properties:
        id:
          type: integer
          description: "Unique ID of the auction"
        gachaCollectionId:
          type: integer
          description: "ID of the associated gacha collection"
        auctionStart:
          type: string
          format: date-time
          description: "Start time of the auction"
        auctionEnd:
          type: string
          format: date-time
          description: "End time of the auction"
        minimumBid:
          type: number
          format: float
          description: "Minimum bid amount for the auction"
        timestamp:
          type: string
          format: date-time
          description: "Timestamp of the auction creation"
        status:
          type: string
          description: "Current status of the auction"
        verified:
          type: boolean
          description: "Indicates whether the auction is verified"
...
    UpdateAuctionRequest:
      type: object
      properties:
        gachaCollectionId:
          type: integer
          description: "ID of the associated gacha collection"
        auctionStart:
          type: string
          format: date-time
          description: "Start time of the auction"
        auctionEnd:
          type: string
          format: date-time
          description: "End time of the auction"
        minimumBid:
          type: number
          format: float
          description: "Minimum bid amount for the auction"
        timestamp:
          type: string
          format: date-time
          description: "Timestamp of the auction update"
        status:
          type: string
          description: "Updated status of the auction" 
...
    AuctionBid:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier for the auction bid.
        userId:
          type: integer
          description: Identifier for the user who placed the bid.
        bidAmount:
          type: number
          format: float
          description: Amount of the bid.
        auctionId:
          type: integer
          description: Identifier for the associated auction.
        timestamp:
          type: string
          format: date-time
          description: Timestamp for when the bid was placed.
security:
  - bearerAuth: []